MetricsListener improvements
- Fix exception on shutdown :
when we close the HttpListener that backs the MetricsListener, the
blocking call to GetContext is interrupted and throws an exception.
Adding a catch block to block it and not let it propagate through.

- Remove a case of tight coupling :
remove the tight coupling between the metrics listener and the
metrics it exposes. Introduced a JsonReportFormatter, that does exactly
that so that it's possible to provide the listener with a different
set of metrics to expose. Modifed HumanReadableReportFormatter
accordingly.